insert boldface into MathML descendant of strong: https://github.com/metanorma/metanorma-iso/issues/876
