+++ spec/isodoc/presentation_xml_spec.rb
@@ -1902,10 +1902,92 @@ RSpec.describe IsoDoc do
       .to be_equivalent_to xmlpp(presxml)
   end
 
+  it "propagates boldface into MathML" do
+    FileUtils.rm_f "test.doc"
+    FileUtils.rm_f "test.html"
+    input = <<~INPUT
+      <iso-standard xmlns="http://riboseinc.com/isoxml">
+                <preface><foreword>
+                <note>
+                <strong><stem type="MathML">
+        <math xmlns="http://www.w3.org/1998/Math/MathML">
+          <mfenced open="[" close="]">
+            <mrow>
+              <mi>a</mi>
+              <mo>,</mo>
+              <mi>b</mi>
+            </mrow>
+          </mfenced>
+        </math>
+        <asciimath>[a,b]</asciimath>
+      </stem></strong>
+      <stem type="MathML">
+        <math xmlns="http://www.w3.org/1998/Math/MathML">
+          <mfenced open="[" close="]">
+            <mrow>
+              <mi>a</mi>
+              <mo>,</mo>
+              <mi>b</mi>
+            </mrow>
+          </mfenced>
+        </math>
+        <asciimath>[a,b]</asciimath>
+      </stem>
+            </note>
+                </foreword></preface>
+      </iso-standard>
+    INPUT
+    presxml = <<~OUTPUT
+          <iso-standard xmlns="http://riboseinc.com/isoxml" type="presentation">
+        <preface>
+          <foreword displayorder="1">
+            <note>
+              <name>NOTE</name>
+              <strong>
+                <stem type="MathML">
+                  <math xmlns="http://www.w3.org/1998/Math/MathML">
+                    <mstyle fontweight="bold">
+                      <mfenced open="[" close="]">
+                        <mrow>
+                          <mi>a</mi>
+                          <mo>,</mo>
+                          <mi>b</mi>
+                        </mrow>
+                      </mfenced>
+                    </mstyle>
+                  </math>
+                  <!-- [a , b] -->
+                  <asciimath>[a,b]</asciimath>
+                </stem>
+              </strong>
+              <stem type="MathML">
+                <math xmlns="http://www.w3.org/1998/Math/MathML">
+                  <mfenced open="[" close="]">
+                    <mrow>
+                      <mi>a</mi>
+                      <mo>,</mo>
+                      <mi>b</mi>
+                    </mrow>
+                  </mfenced>
+                </math>
+                <!-- [a , b] -->
+                <asciimath>[a,b]</asciimath>
+              </stem>
+            </note>
+          </foreword>
+        </preface>
+      </iso-standard>
+    OUTPUT
+    expect(xmlpp(IsoDoc::PresentationXMLConvert.new({})
+      .convert("test", input, true))
+      .sub(%r{<localized-strings>.*</localized-strings>}m, ""))
+      .to be_equivalent_to xmlpp(presxml)
+  end
+
   private
 
   def mock_symbols
-    allow_any_instance_of(::IsoDoc::PresentationXMLConvert)
+    allow_any_instance_of(IsoDoc::PresentationXMLConvert)
       .to receive(:twitter_cldr_localiser_symbols).and_return(group: "'")
   end
 end
